requirements.txt
# ── Core ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Data pipeline ─────────────────────────────────────────
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0
python-docx==1.1.2

# ── GCP / Vertex AI ───────────────────────────────────────
google-cloud-aiplatform==1.71.1
google-cloud-storage==2.16.0
vertexai==1.71.1
google-auth==2.29.0

# ── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.1
tenacity==8.3.0
tqdm==4.66.4

# ── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0